+++ lib/resty/mlcache.lua
@@ -174,7 +174,7 @@ end
 
 
 local _M     = {
-    _VERSION = "2.6.1",
+    _VERSION = "2.7.0",
     _AUTHOR  = "Thibault Charbonnier",
     _LICENSE = "MIT",
     _URL     = "https://github.com/thibaultcha/lua-resty-mlcache",
@@ -616,6 +616,7 @@ local function check_opts(self, opts)
     local resurrect_ttl
     local l1_serializer
     local shm_set_tries
+    local resty_lock_opts
 
     if opts ~= nil then
         if type(opts) ~= "table" then
@@ -670,6 +671,13 @@ local function check_opts(self, opts)
                 error("opts.shm_set_tries must be >= 1", 3)
             end
         end
+
+        resty_lock_opts = opts.resty_lock_opts
+        if resty_lock_opts ~= nil then
+            if type(resty_lock_opts) ~= "table" then
+                error("opts.resty_lock_opts must be a table", 3)
+            end
+        end
     end
 
     if not ttl then
@@ -692,7 +700,12 @@ local function check_opts(self, opts)
         shm_set_tries = self.shm_set_tries
     end
 
-    return ttl, neg_ttl, resurrect_ttl, l1_serializer, shm_set_tries
+    if not resty_lock_opts then
+        resty_lock_opts = self.resty_lock_opts
+    end
+
+    return ttl, neg_ttl, resurrect_ttl, l1_serializer, shm_set_tries,
+           resty_lock_opts
 end
 
 
@@ -707,9 +720,9 @@ end
 
 
 local function run_callback(self, key, shm_key, data, ttl, neg_ttl,
-    went_stale, l1_serializer, resurrect_ttl, shm_set_tries, cb, ...)
+    went_stale, l1_serializer, resurrect_ttl, shm_set_tries, rlock_opts, cb, ...)
 
-    local lock, err = resty_lock:new(self.shm_locks, self.resty_lock_opts)
+    local lock, err = resty_lock:new(self.shm_locks, rlock_opts)
     if not lock then
         return nil, "could not create lock: " .. err
     end
@@ -877,8 +890,8 @@ function _M:get(key, opts, cb, ...)
 
     -- opts validation
 
-    local ttl, neg_ttl, resurrect_ttl, l1_serializer, shm_set_tries =
-        check_opts(self, opts)
+    local ttl, neg_ttl, resurrect_ttl, l1_serializer, shm_set_tries,
+          rlock_opts = check_opts(self, opts)
 
     local err, went_stale, is_stale
     data, err, went_stale, is_stale = get_shm_set_lru(self, key, namespaced_key,
@@ -906,7 +919,7 @@ function _M:get(key, opts, cb, ...)
 
     return run_callback(self, key, namespaced_key, data, ttl, neg_ttl,
                         went_stale, l1_serializer, resurrect_ttl,
-                        shm_set_tries, cb, ...)
+                        shm_set_tries, rlock_opts, cb, ...)
 end
 
 
@@ -922,6 +935,7 @@ local function run_thread(self, ops, from, to)
                                                       ctx.l1_serializer,
                                                       ctx.resurrect_ttl,
                                                       ctx.shm_set_tries,
+                                                      ctx.rlock_opts,
                                                       ctx.cb, ctx.arg)
     end
 end
@@ -1042,8 +1056,8 @@ function _M:get_bulk(bulk, opts)
             res[res_idx + 2] = 1
 
         else
-            local pok, ttl, neg_ttl, resurrect_ttl, l1_serializer, shm_set_tries
-                = pcall(check_opts, self, b_opts)
+            local pok, ttl, neg_ttl, resurrect_ttl, l1_serializer,
+                  shm_set_tries, rlock_opts = pcall(check_opts, self, b_opts)
             if not pok then
                 -- strip the stacktrace
                 local err = ttl:match("mlcache%.lua:%d+:%s(.*)")
@@ -1096,6 +1110,7 @@ function _M:get_bulk(bulk, opts)
                 ctx.l1_serializer = l1_serializer
                 ctx.resurrect_ttl = resurrect_ttl
                 ctx.shm_set_tries = shm_set_tries
+                ctx.rlock_opts = rlock_opts
                 ctx.data = data
                 ctx.err = nil
                 ctx.hit_lvl = nil
@@ -1262,7 +1277,16 @@ function _M:peek(key, stale)
                         "retrieval: " .. err
         end
 
-        local remaining_ttl = ttl - (now() - at)
+        local remaining_ttl = 0
+
+        if ttl > 0 then
+            remaining_ttl = ttl - (now() - at)
+
+            if remaining_ttl == 0 then
+                -- guarantee a non-zero remaining_ttl if ttl is set
+                remaining_ttl = 0.001
+            end
+        end
 
         return remaining_ttl, nil, value, went_stale
     end